refactor: moved vector parent components under vector folder
