+++ sites/geohub/src/components/maplibre/raster/Hillshade.svelte
