+++ sites/geohub/src/components/maplibre/vector/VectorCircle.svelte
