+++ sites/geohub/src/components/maplibre/vector/VectorFillExtrusion.svelte
