+++ sites/geohub/src/components/maplibre/vector/VectorHeatmap.svelte
