+++ sites/geohub/src/components/maplibre/vector/VectorPolygon.svelte
